Break out npc dialogue
